requirements.txt
fastapi>=0.115.0
uvicorn>=0.31.0
python-multipart>=0.0.18
sqlite-vec>=0.1.0
numpy>=2.1.0
sentence-transformers>=3.0.0
pypdf>=4.0.0
python-docx>=1.1.0
jieba>=0.42.1
httpx>=0.27.0